requirements.txt
flask-restful
flask
PyJWT
google-api-python-client
